hist: rewrite to be more "large-ish data" friendly
The new version no longer keeps all the individual items in memory, but
assumes that we're working with a relatively small universe such that we
can keep counts of all the distinct keys in a map.  This makes it
possible to process much much larger data sets.

We now also print percentiles (configurable) and have switched to using
github.com/dgryski/go-onlinestats for the stats calculations.