intset: add grouped varint
http://www.ir.uwaterloo.ca/book/addenda-06-index-compression.html

In order to see an improvement over regular varint, there needs to be a
larger variability in the size of the deltas.  At the current 'quality'
level, all the deltas fit into a single byte and so varint suffers no
branch-prediction problems.  Dropping quality to 0.0001 or smaller gives
grvint the lead.