interp: fix search
Almost all the implementations of interpolation search match the one
given on the wikipedia page, which has overflow issues.  This one comes
from Voldemort, from a blog posting on LinkedIn's tech blog.

I've updated the code to match the semantics of sort.SearchInts() so it
can be used as a drop-in replacement.