matcher: use []byte or string, whichever is 'native'
For some of the matchers we were incurring extra []byte->string
conversion cost.  This stores an extra copy of the input so that we can
see just the effect of the lookup algorithm and not this overhead.

You could argue that this is more artificial, since in the real world
we're not going to keep two copies of the data around.  However, in some
situations you have string, and in some []byte, so lets present the
options in their most positive light.